Build mods to share explciit disclosure testing code
