+++ daml-lf/interpreter/BUILD.bazel
@@ -47,6 +47,9 @@ da_scala_library(
 da_scala_library(
     name = "interpreter-test-lib",
     srcs = glob(["src/test/**/*Lib.scala"]),
+    scala_deps = [
+        "@maven//:org_scalatest_scalatest_matchers_core",
+    ],
     scalacopts = lf_scalacopts_stricter,
     deps = [
         ":interpreter",
